requirements.txt
streamlit==1.46.0
ultralytics==8.0.196
pillow==10.0.1
numpy==1.24.3
torch==2.5.0
torchvision>=0.14.0
gunicorn==21.2.0 